+++ ui/Dashboard.tsx
@@ -1,1 +1,5 @@
-export function Dashboard() {}
+import React from 'react';
+
+export function Dashboard() {
+  return <span>Dashboard</span>;
+}
