+++ ui/components/Loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export function Loading() {
+  return (
+    <div className="loading">
+      Loading...
+      <span id="spin"></span>
+    </div>
+  );
+}
